Moved NetworkUtil to networking package, last commit before implementing sliding up panel for queued songs.
